@@ -15,75 +15,55 @@
     "db:psql": "node ./db/connect.js",
     "electric:start": "node ./backend/startElectric.js",
     "prepare-wa-sqlite": "cp ./node_modules/wa-sqlite/dist/wa-sqlite-async.mjs ./public/wa-sqlite-async.mjs && cp ./node_modules/wa-sqlite/dist/wa-sqlite-async.wasm ./public/wa-sqlite-async.wasm",
-    "OLD-migrate": "node ./db-OLD/migrate.tsx",
-    "OLD-data": "node ./db-OLD/data.tsx",
-    "build": "npm run prepare-wa-sqlite && rm -rf dist && node builder.js",
-    "build-prod": "rm -rf dist && NODE_ENV=production node builder.js",
-    "start": "SERVE=true node builder.js"
-  },
-  "engines": {
-    "node": ">=16.11.0"
+    "dev": "vite",
+    "start": "npm run dev",
+    "build": "npm run prepare-wa-sqlite && tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@headlessui/react": "^0.3.2",
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "@prisma/client": "4.12.0",
-    "@tailwindcss/forms": "^0.3.2",
-    "@tailwindcss/line-clamp": "^0.2.0",
+    "@firefox-devtools/react-contextmenu": "^5.1.0",
+    "@headlessui/react": "^1.7.17",
+    "@prisma/client": "^4.12.0",
+    "@svgr/plugin-jsx": "^8.1.0",
+    "@svgr/plugin-svgo": "^8.1.0",
+    "@tailwindcss/forms": "^0.5.6",
     "animate.css": "^4.1.1",
-    "axios": "^1.3.5",
     "classnames": "^2.3.2",
-    "dayjs": "^1.10.4",
+    "dayjs": "^1.11.9",
     "electric-sql": "latest",
-    "fs-extra": "^10.1.0",
-    "jsonwebtoken": "^9.0.0",
-    "react": ">= 17.0.2",
-    "react-beautiful-dnd": "^13.1.0",
-    "react-contexify": "^5.0.0",
-    "react-contextmenu": "^2.14.0",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.2.0",
-    "react-popper": "^2.2.5",
-    "react-portal": "^4.2.1",
-    "react-redux": "^7.2.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-toastify": "^7.0.3",
-    "redux": "^4.0.5",
-    "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.3.0",
-    "rich-markdown-editor": "^11.21.3",
-    "styled-components": "^5.3.5",
+    "jsonwebtoken": "^9.0.2",
+    "react": "^18.2.0",
+    "react-beautiful-dnd": "^13.1.1",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.10.1",
+    "react-router-dom": "^6.15.0",
+    "react-toastify": "^9.1.3",
     "uuid": "^9.0.0",
-    "wa-sqlite": "rhashimoto/wa-sqlite#master",
-    "zod": "^3.20.2"
+    "vite-plugin-svgr": "^3.2.0",
+    "wa-sqlite": "github:rhashimoto/wa-sqlite#master",
+    "zod": "^3.22.2"
   },
   "devDependencies": {
     "@databases/pg": "^5.4.1",
     "@electric-sql/prisma-generator": "latest",
-    "@tailwindcss/postcss7-compat": "^2.1.0",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^16.18.16",
-    "@types/react": "^17.0.2",
-    "@types/react-beautiful-dnd": "^13.0.0",
-    "@types/react-dom": "^17.0.2",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/uuid": "^9.0.1",
-    "autoprefixer": "^9.8.6",
-    "esbuild": "^0.16.17",
-    "esbuild-css-modules-plugin": "^2.7.1",
-    "esbuild-plugin-inline-image": "^0.0.9",
-    "esbuild-style-plugin": "^1.6.2",
-    "esbuild-svgr-plugin": "^0.2.0",
-    "postcss": "^7.0.35",
-    "prisma": "4.12.0",
-    "react-dom": "^17.0.2",
-    "shelljs": "^0.8.5",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.1.0",
-    "typescript": "^4.5.5"
-  },
-  "overrides": {
-    "prosemirror-tables": "^1.1.1"
+    "@types/jest": "^29.5.4",
+    "@types/node": "^20.5.9",
+    "@types/react": "^18.2.15",
+    "@types/react-beautiful-dnd": "^13.1.4",
+    "@types/react-dom": "^18.2.7",
+    "@types/react-router-dom": "^5.3.3",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "@vitejs/plugin-react": "^4.0.3",
+    "autoprefixer": "^10.4.15",
+    "eslint": "^8.45.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "postcss": "^8.4.29",
+    "prisma": "^4.12.0",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.0.2",
+    "vite": "^4.4.5"
   }
 }
