@@ -1,13 +1,14 @@
 
 import React from 'react';
-import ReactDOM from 'react-dom';
-// import { Provider } from 'react-redux';
+import { createRoot } from 'react-dom/client';
 import './style.css';
 
 import App from './App';
 
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const container = document.getElementById('root')!;
+const root = createRoot(container);
+root.render(<App />);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
