@@ -1,10 +1,6 @@
-const colors = require('tailwindcss/colors');
-
-
-module.exports = {
-  // mode: "jit",
-  purge: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
-  darkMode: false, // or 'media' or 'class'
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
   theme: {
     screens: {
 
@@ -76,7 +72,7 @@ module.exports = {
     },
   },
   plugins: [
-    require('@tailwindcss/line-clamp'),
     require('@tailwindcss/forms')
   ],
 }
+
