@@ -1,11 +1,11 @@
-import AddIcon from '../assets/icons/add.svg';
-import ArchiveIcon from '../assets/icons/archive.svg';
-import HelpIcon from '../assets/icons/help.svg';
-import InboxIcon from '../assets/icons/inbox.svg';
-import IssueIcon from '../assets/icons/issue.svg';
-import MenuIcon from '../assets/icons/menu.svg';
-import ProjectIcon from '../assets/icons/project.svg';
-import ViewIcon from '../assets/icons/view.svg';
+import { ReactComponent as AddIcon } from '../assets/icons/add.svg';
+import { ReactComponent as ArchiveIcon } from '../assets/icons/archive.svg';
+import { ReactComponent as HelpIcon } from '../assets/icons/help.svg';
+import { ReactComponent as InboxIcon } from '../assets/icons/inbox.svg';
+import { ReactComponent as IssueIcon } from '../assets/icons/issue.svg';
+import { ReactComponent as MenuIcon } from '../assets/icons/menu.svg';
+import { ReactComponent as ProjectIcon } from '../assets/icons/project.svg';
+import { ReactComponent as ViewIcon } from '../assets/icons/view.svg';
 import classnames from 'classnames';
 import SearchBox from '../components/SearchBox';
 import { useClickOutside } from '../hooks/useClickOutside';
