@@ -1,8 +1,8 @@
-import SignalUrgentIcon from '../assets/icons/claim.svg';
-import SignalNoPriorityIcon from '../assets/icons/dots.svg';
-import SignalMediumIcon from '../assets/icons/signal-medium.svg';
-import SignalStrongIcon from '../assets/icons/signal-strong.svg';
-import SignalWeakIcon from '../assets/icons/signal-weak.svg';
+import { ReactComponent as SignalUrgentIcon } from '../assets/icons/claim.svg';
+import { ReactComponent as SignalNoPriorityIcon } from '../assets/icons/dots.svg';
+import { ReactComponent as SignalMediumIcon } from '../assets/icons/signal-medium.svg';
+import { ReactComponent as SignalStrongIcon } from '../assets/icons/signal-strong.svg';
+import { ReactComponent as SignalWeakIcon } from '../assets/icons/signal-weak.svg';
 import classNames from 'classnames';
 import React from 'react';
 import { Priority } from '../types/issue';
